requirements.txt
streamlit
transformers
torch
faiss-cpu
sentence-transformers
python-docx
reportlab
requests
toml
